refactor: move extensions to member package
