+++ src/main/kotlin/nz/adjmunro/knomadic/fetch/flow/FlowFlatten.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import nz.adjmunro.knomadic.FetchFlow
 import nz.adjmunro.knomadic.fetch.Fetch
-import nz.adjmunro.knomadic.fetch.flatten
+import nz.adjmunro.knomadic.fetch.members.flatten
 
 /** Convenience function to apply [flatten] inside a [Flow]. */
 public fun <T : Any> FetchFlow<Fetch<T>>.flatten(): FetchFlow<T> {
