+++ src/test/kotlin/nz/adjmunro/knomadic/fetch/SuspendTest.kt
@@ -19,6 +19,29 @@ import nz.adjmunro.knomadic.fetch.flow.mapFinished
 import nz.adjmunro.knomadic.fetch.flow.mapInitialToFinished
 import nz.adjmunro.knomadic.fetch.flow.mapToFinished
 import nz.adjmunro.knomadic.fetch.flow.onEachFetch
+import nz.adjmunro.knomadic.fetch.members.flatMapFetching
+import nz.adjmunro.knomadic.fetch.members.flatMapFinished
+import nz.adjmunro.knomadic.fetch.members.flatMapInitial
+import nz.adjmunro.knomadic.fetch.members.flatten
+import nz.adjmunro.knomadic.fetch.members.fold
+import nz.adjmunro.knomadic.fetch.members.getOrDefault
+import nz.adjmunro.knomadic.fetch.members.getOrElse
+import nz.adjmunro.knomadic.fetch.members.getOrNull
+import nz.adjmunro.knomadic.fetch.members.getOrThrow
+import nz.adjmunro.knomadic.fetch.members.isFetching
+import nz.adjmunro.knomadic.fetch.members.isFinished
+import nz.adjmunro.knomadic.fetch.members.isNotFetching
+import nz.adjmunro.knomadic.fetch.members.isNotFinished
+import nz.adjmunro.knomadic.fetch.members.isNotStarted
+import nz.adjmunro.knomadic.fetch.members.isStarted
+import nz.adjmunro.knomadic.fetch.members.mapFetchingToFinished
+import nz.adjmunro.knomadic.fetch.members.mapFinished
+import nz.adjmunro.knomadic.fetch.members.mapInitialToFinished
+import nz.adjmunro.knomadic.fetch.members.mapToFinished
+import nz.adjmunro.knomadic.fetch.members.onFetching
+import nz.adjmunro.knomadic.fetch.members.onFinished
+import nz.adjmunro.knomadic.fetch.members.onInitial
+import nz.adjmunro.knomadic.fetch.members.unwrap
 import kotlin.test.Ignore
 import kotlin.test.Test
