+++ src/test/kotlin/nz/adjmunro/knomadic/outcome/SuspendTest.kt
@@ -6,6 +6,26 @@ import io.kotest.matchers.result.shouldBeSuccess
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.test.TestResult
 import kotlinx.coroutines.test.runTest
+import nz.adjmunro.knomadic.outcome.members.coerceToFailure
+import nz.adjmunro.knomadic.outcome.members.coerceToSuccess
+import nz.adjmunro.knomadic.outcome.members.errorOrElse
+import nz.adjmunro.knomadic.outcome.members.falter
+import nz.adjmunro.knomadic.outcome.members.flatMapFailure
+import nz.adjmunro.knomadic.outcome.members.flatMapSuccess
+import nz.adjmunro.knomadic.outcome.members.fold
+import nz.adjmunro.knomadic.outcome.members.getOrElse
+import nz.adjmunro.knomadic.outcome.members.isFailure
+import nz.adjmunro.knomadic.outcome.members.isSuccess
+import nz.adjmunro.knomadic.outcome.members.map
+import nz.adjmunro.knomadic.outcome.members.mapEachFailure
+import nz.adjmunro.knomadic.outcome.members.mapEachSuccess
+import nz.adjmunro.knomadic.outcome.members.mapFailure
+import nz.adjmunro.knomadic.outcome.members.mapSuccess
+import nz.adjmunro.knomadic.outcome.members.onEachFailure
+import nz.adjmunro.knomadic.outcome.members.onEachSuccess
+import nz.adjmunro.knomadic.outcome.members.onFailure
+import nz.adjmunro.knomadic.outcome.members.onSuccess
+import nz.adjmunro.knomadic.outcome.members.recover
 import nz.adjmunro.knomadic.raise.RaiseScope
 import nz.adjmunro.knomadic.raise.RaiseScope.Companion.catch
 import nz.adjmunro.knomadic.raise.RaiseScope.Companion.default
